Compiler warning fixes
--HG--
branch : HEAD

+++ src/lib-storage/mailbox-uidvalidity.c
@@ -22,8 +22,8 @@ static uint32_t mailbox_uidvalidity_next_fallback(void)
 	/* we failed to use the uidvalidity file. don't fail the mailbox
 	   creation because of it though, most of the time it's safe enough
 	   to use the current time as the uidvalidity value. */
-	if (uid_validity < ioloop_time)
-		uid_validity = ioloop_time;
+	if (uid_validity < (uint32_t)ioloop_time)
+		uid_validity = (uint32_t)ioloop_time;
 	else
 		uid_validity++;
 	return uid_validity;
